+++ R/app_server.R
@@ -63,7 +63,7 @@ app_server <- function(input, output, session) {
 
 
     if (isTRUE(global$missingdata)) {
-      info <- get_info_oak_last_truffle(dbtruffle = global$truffe, theidoak = input$chene_click, filter_missing_info = TRUE)
+      info <- get_info_oak_truffle(dbtruffle = global$truffe, theidoak = input$chene_click, filter_missing_info = TRUE)
 
       golem::invoke_js(
         "modal_info_missing",
@@ -91,7 +91,7 @@ app_server <- function(input, output, session) {
           type = info$chene$type,
           date_reens = info$reensemence,
           date_p = as.Date(info$chene$planting_date),
-          der_truf = info$truffes$derniere_truffe,
+          der_truf = info$truffes$last_truffle,
           tot_weight = info$truffes$weight_tot,
           last_comment = info$truffes$last_comment,
           other_comments = info$truffes$other_comments