+++ tests/testthat/test-get_info.R
@@ -38,7 +38,7 @@ test_that("get_info works", {
     ),
     truffes = list(
       weight_tot = 0,
-      derniere_truffe = "-",
+      last_truffle = "-",
       last_comment = "-",
       other_comments = "-"
     ),
@@ -56,7 +56,7 @@ test_that("get_info works", {
       ),
       truffes = list(
         weight_tot = 125,
-        derniere_truffe = structure(19844, class = "Date"),
+        last_truffle = structure(19844, class = "Date"),
         last_comment = "2024-05-01 : Medium truffle",
         other_comments = "2024-04-01 : Small truffle"
       ),